+++ adminer/drivers/oracle.inc.php
@@ -391,7 +391,6 @@ ORDER BY ac.constraint_type, aic.column_position", $connection2) as $row
 		$queries = array();
 		foreach ($alter as $val) {
 			if ($val[0] != "INDEX") {
-				//! descending UNIQUE indexes results in syntax error
 				$val[2] = preg_replace('~ DESC$~', '', $val[2]);
 				$create = ($val[2] == "DROP"
 					? "\nDROP CONSTRAINT " . idf_escape($val[1])