+++ adminer/lang/cs.inc.php
@@ -210,6 +210,7 @@ Lang::$translations = array(
 	'Add next' => 'Přidat další',
 	'Index Type' => 'Typ indexu',
 	'length' => 'délka',
+	'Algorithm' => 'Algoritmus',
 
 	'Foreign keys' => 'Cizí klíče',
 	'Foreign key' => 'Cizí klíč',
@@ -355,7 +356,6 @@ Lang::$translations = array(
 	'%s must <a%s>return an array</a>.' => '%s musí <a%s>vracet pole</a>.',
 	'<a%s>Configure</a> %s in %s.' => '<a%s>Nakonfigurujte</a> %s v %s.',
 	'screenshot' => 'obrázek',
-	'Algorithm' => 'Algoritmus',
 );
 
 // run `php ../../lang.php cs` to update this file