+++ adminer/select.inc.php
@@ -389,10 +389,12 @@ if (!$columns && support("table")) {
 				$unique_array = unique_array($rows[$n], $indexes);
 				if (!$unique_array) {
 					$unique_array = array();
+					reset($select);
 					foreach ($rows[$n] as $key => $val) {
-						if (!preg_match('~^(COUNT\((\*|(DISTINCT )?`(?:[^`]|``)+`)\)|(AVG|GROUP_CONCAT|MAX|MIN|SUM)\(`(?:[^`]|``)+`\))$~', $key)) { //! columns looking like functions
+						if (!preg_match('~^(COUNT|AVG|GROUP_CONCAT|MAX|MIN|SUM)\(~', current($select))) {
 							$unique_array[$key] = $val;
 						}
+						next($select);
 					}
 				}
 				$unique_idf = "";
@@ -410,8 +412,10 @@ if (!$columns && support("table")) {
 					. ($is_group || information_schema(DB) ? "" : " <a href='" . h(ME . "edit=" . urlencode($TABLE) . $unique_idf) . "' class='edit'>" . lang('edit') . "</a>")
 				);
 
+				reset($select);
 				foreach ($row as $key => $val) {
 					if (isset($names[$key])) {
+						$column = current($select);
 						$field = (array) $fields[$key];
 						$val = driver()->value($val, $field);
 						if ($val != "" && (!isset($email_fields[$key]) || $email_fields[$key] != "")) {
@@ -439,7 +443,7 @@ if (!$columns && support("table")) {
 								}
 							}
 						}
-						if ($key == "COUNT(*)") { //! columns looking like functions
+						if ($column == "COUNT(*)") {
 							$link = ME . "select=" . urlencode($TABLE);
 							$i = 0;
 							foreach ((array) $_GET["where"] as $v) {
@@ -457,7 +461,11 @@ if (!$columns && support("table")) {
 						$posted = idx(idx($_POST["val"], $unique_idf), bracket_escape($key));
 						$editable = !is_array($row[$key]) && is_utf8($html) && $rows[$n][$key] == $row[$key] && !$functions[$key] && !$field["generated"];
 						$text = preg_match('~text|json|lob~', $field["type"]);
-						echo "<td id='$id'" . (preg_match(number_type(), $field["type"]) && ($val === null || is_numeric(strip_tags($html))) ? " class='number'" : "");
+						$is_number = preg_match(number_type(), $field["type"])
+							|| preg_match('~^(CHAR_LENGTH|ROUND|FLOOR|CEIL|TIME_TO_SEC|COUNT|SUM)\(~', $column)
+							|| (preg_match('~^(AVG|MIN|MAX)\((.+)\)~', $column, $match) && preg_match(number_type(), $fields[idf_unescape($match[2])]["type"]))
+						;
+						echo "<td id='$id'" . ($is_number && ($val === null || is_numeric(strip_tags($html))) ? " class='number'" : "");
 						if (($_GET["modify"] && $editable && $val !== null) || $posted !== null) {
 							$h_value = h($posted !== null ? $posted : $row[$key]);
 							echo ">" . ($text ? "<textarea name='$id' cols='30' rows='" . (substr_count($row[$key], "\n") + 1) . "'>$h_value</textarea>" : "<input name='$id' value='$h_value' size='$lengths[$key]'>");
@@ -469,6 +477,7 @@ if (!$columns && support("table")) {
 							;
 						}
 					}
+					next($select);
 				}
 
 				if ($backward_keys) {