+++ plugins/table-indexes-structure.php
@@ -8,14 +8,11 @@
 */
 class AdminerTableIndexesStructure extends Adminer\Plugin {
 
-	/** Print table structure in tabular format
-	* @param Index[] $indexes data about all indexes on a table
-	*/
-	function tableIndexesPrint($indexes): bool {
+	function tableIndexesPrint($indexes, $tableStatus): bool {
 		echo "<table>\n";
-		echo "<thead><tr><th>" . Adminer\lang('Name') . "<th>" . Adminer\lang('Type') . "<th>" . Adminer\lang('Columns') . "</thead>\n";
+		echo "<thead><tr><th>" . Adminer\lang('Name') . "<th>" . Adminer\lang('Type') . "<th>" . Adminer\lang('Algorithm') . "<th>" . Adminer\lang('Columns') . "</thead>\n";
 		foreach ($indexes as $name => $index) {
-			echo "<tr><th>" . Adminer\h($name) . "<td>" . $index['type'];
+			echo "<tr><th>" . Adminer\h($name) . "<td>$index[type]<td>$index[algorithm]";
 			ksort($index["columns"]); // enforce correct columns order
 			$print = array();
 			foreach ($index["columns"] as $key => $val) {