+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 ## Adminer dev
+- Align numeric functions right
+- Autocomplete: Support table aliases
+- Fix type error in Create function (bug #1053)
 - MySQL: Avoid warning on selecting tables with fulltext indexes (bug #1036)
+- MySQL, PostgreSQL: Support index algorithms (bug #1030)
 - PostgreSQL, CockroachDB: Creating partitioned tables (bug #1031)
 - PostgreSQL: Move partitioned tables from table list to parent table
-- PostgreSQL: Support index algorithms (bug #1030)
 - PostgreSQL: Support partial indices (bug #1048)
 - PostgreSQL: Support calling functions returning table (bug #1040)
-- Designs: adminer.css with 'prefers-color-scheme: dark' don't disable dark mode
+- Designs: adminer.css with 'prefers-color-scheme: dark' doesn't disable dark mode
 - Plugins: Method bodyClass() to add &lt;body class>
+- Plugins: Allow setting dark mode in css()
+- Hindi translation
 
 ## Adminer 5.2.1 (released 2025-04-11)
 - Fix search anywhere (bug #1004, regression from 5.1.1)